backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
numpy==2.1.1
scipy==1.14.1
pytest==8.3.3
# OCC bindings for production geometry ops
pythonocc-core==7.7.2
